Allow @mutt types to derive from abstract supertypes via traits.
Make Mutts types implement a trait (`mutts_trait `) instead of deriving
from a common abstract type. This allows `@mutt struct S <: Base ...  end`
to work, since the Muttsiness of a type doesn't get in the way of its
supertype declaration. :)

In order to keep `setproperty!` working, we also have to generate a new
overload for `setproperty!()` for each type, which is disappointing but
oh well.

Also, in this PR I removed the overload of Base.isimmutable because
after thinking more about it, I think it's more about whether a type is
implemented as a _julia_ "immutable struct" (ie it logically doesn't
have an address). I don't want to "lie" about that. We can always add
this back in if we change our minds. :)